fix: include dpr parameter in fixed-width mode, pull target_width calculations out as a constant
