+++ imgix/urlbuilder.py
@@ -4,12 +4,11 @@ import re
 
 from .urlhelper import UrlHelper
 
-from .constants import DOMAIN_PATTERN
+from .constants import DOMAIN_PATTERN, _target_widths
 
 
-SRCSET_INCREMENT_PERCENTAGE = 8
-SRCSET_MAX_SIZE = 8192
 SRCSET_DPR_TARGET_RATIOS = range(1, 6)
+SRCSET_TARGET_WIDTHS = _target_widths()
 
 
 class UrlBuilder(object):
@@ -151,26 +150,11 @@ class UrlBuilder(object):
         else:
             return self._build_srcset_pairs(path, params)
 
-    def _target_widths(self):
-        resolutions = []
-        prev = 100
-
-        def ensure_even(n):
-            return 2 * round(n/2.0)
-
-        while prev <= SRCSET_MAX_SIZE:
-            resolutions.append(int(ensure_even(prev)))
-            prev *= 1 + (SRCSET_INCREMENT_PERCENTAGE / 100.0) * 2
-
-        resolutions.append(SRCSET_MAX_SIZE)
-        return resolutions
-
     def _build_srcset_pairs(self, path, params):
         srcset = ''
-        widths = self._target_widths()
 
-        for i in range(len(widths)):
-            current_width = widths[i]
+        for i in range(len(SRCSET_TARGET_WIDTHS)):
+            current_width = SRCSET_TARGET_WIDTHS[i]
             current_params = params
             current_params['w'] = current_width
             srcset += self.create_url(path, current_params) \
@@ -180,10 +164,12 @@ class UrlBuilder(object):
 
     def _build_srcset_DPR(self, path, params):
         srcset = ''
-        url = self.create_url(path, params)
 
         for i in range(len(SRCSET_DPR_TARGET_RATIOS)):
             current_ratio = SRCSET_DPR_TARGET_RATIOS[i]
-            srcset += url + ' ' + str(current_ratio) + 'x,\n'
+            current_params = params
+            current_params['dpr'] = i+1
+            srcset += self.create_url(path, current_params) \
+                + ' ' + str(current_ratio) + 'x,\n'
 
         return srcset[0:-2]
